# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rooms": {
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행방 생성",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/from-compatibility": {
      "post": {
        "operationId": "RoomsController_createFromCompatibility",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFromCompatibilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "궁합 멤버 목록으로 여행방 생성",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/me": {
      "get": {
        "operationId": "RoomsController_listMine",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ongoing",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OngoingTripDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 여행방 목록",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}": {
      "get": {
        "operationId": "RoomsController_getRoom",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행방 상세",
        "tags": [
          "여행방"
        ]
      },
      "patch": {
        "operationId": "RoomsController_updateRoom",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행방 수정 (title, dates, status)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/summary": {
      "get": {
        "operationId": "RoomsController_getSummary",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OngoingTripDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행방 요약 (목록용)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/destination": {
      "patch": {
        "operationId": "RoomsController_updateDestination",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행지 설정",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/progress": {
      "get": {
        "operationId": "RoomsController_getProgress",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomProgressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "여행 준비 5단계 진행률",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/invite-link": {
      "get": {
        "operationId": "RoomsController_getInviteLink",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteLinkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "초대 링크 조회",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/invites": {
      "post": {
        "operationId": "RoomsController_regenerateInvite",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteLinkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "초대코드 재발급 (owner만)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/workspace": {
      "get": {
        "operationId": "RoomsController_getWorkspace",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "두리 워크스페이스 (room + candidates + schedule)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/compatibility": {
      "get": {
        "operationId": "RoomsController_getCompatibility",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "멤버 궁합 점수 (N명 pairwise tag 평균)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/match-result": {
      "get": {
        "operationId": "RoomsController_getMatchResult",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "궁합 결과 (compatibility와 동일)",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/schedule-style": {
      "patch": {
        "operationId": "RoomsController_setScheduleStyle",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleStyleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "J/P 일정 스타일 설정",
        "tags": [
          "여행방"
        ]
      }
    },
...
    "/api/v1/rooms/{roomId}/candidates/by-member": {
      "get": {
        "operationId": "RoomsController_candidatesByMember",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "멤버별 후보 장소",
        "tags": [
          "여행방"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/candidates/common": {
      "get": {
        "operationId": "RoomsController_commonCandidates",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "2명 이상 공통 후보",
        "tags": [
          "여행방"
        ]
      }
    },
...
    "/api/v1/rooms/{roomId}/analysis-baseline": {
      "get": {
        "description": "scheduleVersion + factsVersion + 기간·planning",
        "operationId": "RoomsController_analysisBaseline",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "분석/제안용 기준 스냅샷",
        "tags": [
          "여행방"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tourmate API",
    "description": "TripMatch / Tourmate backend REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "제주 여행"
          }
        }
      },
...
      "CreateFromCompatibilityDto": {
        "type": "object",
        "properties": {
          "memberUserIds": {
            "example": [
              "665abc123def456789012346",
              "665abc123def456789012347"
            ],
            "description": "함께 여행할 멤버 userId 목록 (본인 제외 가능)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string",
            "example": "궁합 멤버 제주 여행"
          }
        },
        "required": [
          "memberUserIds"
        ]
      },
...
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "제주 3박4일"
          },
          "startDate": {
            "type": "string",
            "example": "2026-07-10"
          },
          "endDate": {
            "type": "string",
            "example": "2026-07-13"
          },
          "status": {
            "type": "string",
            "enum": [
              "ongoing",
              "completed"
            ],
            "example": "ongoing"
          }
        }
      },
...
    }
  }
}